Completed APOGEE allstar and allvisit donwnload function

+++ astroNN/apogeetools/downloader.py
@@ -35,19 +35,37 @@ def allstar(dr=None):
     HISTORY:
         2017-Oct-09 Henry Leung
     """
+
+    # Check if dr arguement is provided, if none then use default
     if dr is None:
         dr = 14
 
     if dr == 13:
-        url = 'https://data.sdss.org/sas/dr13/apogee/spectro/redux/r6/stars/l30e/l30e.2/allStar-l30e.2.fits'
+        # Check if directory exists
+        fullfilepath = os.path.join(currentdir, 'apogee_dr13\\')
+        if not os.path.exists(fullfilepath):
+            os.makedirs(fullfilepath)
+        filename = 'allStar-l30e.2.fits'
+        fullfilename = os.path.join(fullfilepath, filename)
+        url = 'https://data.sdss.org/sas/dr13/apogee/spectro/redux/r6/stars/l30e/l30e.2/{}'.format(filename)
     elif dr == 14:
-        url = 'https://data.sdss.org/sas/dr14/apogee/spectro/redux/r8/stars/l31c/l31c.2/allStar-l31c.2.fits'
+        fullfilepath = os.path.join(currentdir, 'apogee_dr14\\')
+        # Check if directory exists
+        if not os.path.exists(fullfilepath):
+            os.makedirs(fullfilepath)
+        filename = 'allStar-l31c.2.fits'
+        fullfilename = os.path.join(fullfilepath, filename)
+        url = 'https://data.sdss.org/sas/dr14/apogee/spectro/redux/r8/stars/l31c/l31c.2/{}'.format(filename)
     else:
         raise ValueError('[astroNN.apogeetools.downloader.all_star()] only supports DR13 and DR14 APOGEE')
 
-    with TqdmUpTo(unit='B', unit_scale=True, miniters=1, desc=url.split('/')[-1]) as t:
-        urllib.request.urlretrieve(url, reporthook=t.update_to)
-    print('Downloaded DR{:d} allStar file catalog successfulliy to {}'.format(dr, currentdir))
+    # Check if files exists
+    if not os.path.isfile(os.path.join(fullfilepath, filename)):
+        with TqdmUpTo(unit='B', unit_scale=True, miniters=1, desc=url.split('/')[-1]) as t:
+            urllib.request.urlretrieve(url, reporthook=t.update_to)
+            print('Downloaded DR{:d} allStar file catalog successfully to {}'.format(dr, fullfilename))
+    else:
+        print(fullfilename + ' was found, not downloaded again')
 
     return None
 
@@ -65,15 +83,30 @@ def allvisit(dr=None):
         dr = 14
 
     if dr == 13:
-        url = 'https://data.sdss.org/sas/dr13/apogee/spectro/redux/r6/allVisit-l30e.2.fits'
+        # Check if directory exists
+        fullfilepath = os.path.join(currentdir, 'apogee_allvisited_dr13\\')
+        if not os.path.exists(fullfilepath):
+            os.makedirs(fullfilepath)
+        filename = 'allVisit-l30e.2.fits'
+        fullfilename = os.path.join(fullfilepath, filename)
+        url = 'https://data.sdss.org/sas/dr13/apogee/spectro/redux/r6/{}'.format(filename)
     elif dr == 14:
-        url = 'https://data.sdss.org/sas/dr14/apogee/spectro/redux/r8/allVisit-l31c.2.fits'
+        # Check if directory exists
+        fullfilepath = os.path.join(currentdir, 'apogee_allvisited_dr14\\')
+        if not os.path.exists(fullfilepath):
+            os.makedirs(fullfilepath)
+        filename = 'allVisit-l31c.2.fits'
+        fullfilename = os.path.join(fullfilepath, filename)
+        url = 'https://data.sdss.org/sas/dr14/apogee/spectro/redux/r8/{}'.format(filename)
     else:
         raise ValueError('[astroNN.apogeetools.downloader.all_visit()] only supports DR13 and DR14 APOGEE')
 
-    with TqdmUpTo(unit='B', unit_scale=True, miniters=1, desc=url.split('/')[-1]) as t:
-        urllib.request.urlretrieve(url, reporthook=t.update_to)
-    print('Downloaded DR{:d} allVisit file catalog successfulliy to {}'.format(dr, currentdir))
+    if not os.path.isfile(os.path.join(fullfilepath, filename)):
+        with TqdmUpTo(unit='B', unit_scale=True, miniters=1, desc=url.split('/')[-1]) as t:
+            urllib.request.urlretrieve(url, reporthook=t.update_to)
+            print('Downloaded DR{:d} allVisit file catalog successfully to {}'.format(dr, currentdir))
+    else:
+        print(fullfilename + ' was found, not downloaded again')
 
     return None
